Moved Poll to its own component.
